+++ src/Domain/Item/ItemFactoryInterface.ts
@@ -1,0 +1,6 @@
+import { Item } from './Item'
+import { ItemHash } from './ItemHash'
+
+export interface ItemFactoryInterface {
+  create(userUuid: string, itemHash: ItemHash, userAgent?: string): Item
+}
