+++ src/Domain/Item/ItemService.ts
@@ -12,6 +12,7 @@ import { GetItemsDTO } from './GetItemsDTO'
 import { GetItemsResult } from './GetItemsResult'
 import { Item } from './Item'
 import { ItemConflict } from './ItemConflict'
+import { ItemFactoryInterface } from './ItemFactoryInterface'
 import { ItemHash } from './ItemHash'
 import { ItemQuery } from './ItemQuery'
 import { ItemRepositoryInterface } from './ItemRepositoryInterface'
@@ -27,6 +28,7 @@ export class ItemService implements ItemServiceInterface {
 
   constructor (
     @inject(TYPES.ItemSaveProcessor) private itemSaveProcessor: ItemSaveProcessorInterface,
+    @inject(TYPES.ItemFactory) private itemFactory: ItemFactoryInterface,
     @inject(TYPES.ItemRepository) private itemRepository: ItemRepositoryInterface,
     @inject(TYPES.RevisionService) private revisionService: RevisionServiceInterface,
     @inject(TYPES.DomainEventPublisher) private domainEventPublisher: DomainEventPublisherInterface,
@@ -95,6 +97,9 @@ export class ItemService implements ItemServiceInterface {
       if (!processingResult.passed) {
         if (processingResult.conflict) {
           conflicts.push(processingResult.conflict)
+        }
+        if (processingResult.skipped) {
+          savedItems.push(processingResult.skipped)
         }
 
         continue
@@ -229,48 +234,7 @@ export class ItemService implements ItemServiceInterface {
   }
 
   private async saveNewItem(userUuid: string, itemHash: ItemHash, userAgent?: string): Promise<Item> {
-    const newItem = new Item()
-    newItem.uuid = itemHash.uuid
-    if (itemHash.content) {
-      newItem.content = itemHash.content
-    }
-    newItem.userUuid = userUuid
-    if (itemHash.content_type) {
-      newItem.contentType = itemHash.content_type
-    }
-    if (itemHash.enc_item_key) {
-      newItem.encItemKey = itemHash.enc_item_key
-    }
-    if (itemHash.items_key_id) {
-      newItem.itemsKeyId = itemHash.items_key_id
-    }
-    if (itemHash.duplicate_of) {
-      newItem.duplicateOf = itemHash.duplicate_of
-    }
-    if (itemHash.deleted !== undefined) {
-      newItem.deleted = itemHash.deleted
-    }
-    if (itemHash.auth_hash) {
-      newItem.authHash = itemHash.auth_hash
-    }
-    newItem.lastUserAgent = userAgent ?? null
-
-    const now = this.timer.getTimestampInMicroseconds()
-    const nowDate = this.timer.convertMicrosecondsToDate(now)
-
-    newItem.updatedAtTimestamp = now
-    newItem.updatedAt = nowDate
-
-    newItem.createdAtTimestamp = now
-    newItem.createdAt = nowDate
-
-    if (itemHash.created_at_timestamp) {
-      newItem.createdAtTimestamp = itemHash.created_at_timestamp
-      newItem.createdAt = this.timer.convertMicrosecondsToDate(itemHash.created_at_timestamp)
-    } else if (itemHash.created_at) {
-      newItem.createdAtTimestamp = this.timer.convertStringDateToMicroseconds(itemHash.created_at)
-      newItem.createdAt = this.timer.convertStringDateToDate(itemHash.created_at)
-    }
+    const newItem = this.itemFactory.create(userUuid, itemHash, userAgent)
 
     const savedItem = await this.itemRepository.save(newItem)
