+++ src/Domain/Item/SaveFilter/ItemSaveFilteringResult.ts
@@ -1,6 +1,8 @@
+import { Item } from '../Item'
 import { ItemConflict } from '../ItemConflict'
 
 export type ItemSaveFilteringResult = {
   passed: boolean
   conflict?: ItemConflict
+  skipped?: Item
 }
