+++ src/Domain/Item/SaveProcessor/ItemSaveProcessingResult.ts
@@ -1,6 +1,8 @@
+import { Item } from '../Item'
 import { ItemConflict } from '../ItemConflict'
 
 export type ItemSaveProcessingResult = {
   passed: boolean
   conflict?: ItemConflict
+  skipped?: Item
 }
